# backend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TopTrainers API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/exercises": {
      "get": {
        "tags": [
          "exercises"
        ],
        "summary": "List Exercises",
        "operationId": "list_exercises_api_v1_exercises_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExerciseResponse"
                  },
                  "type": "array",
                  "title": "Response List Exercises Api V1 Exercises Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "exercises"
        ],
        "summary": "Create Exercise",
        "operationId": "create_exercise_api_v1_exercises_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/exercises/video-uploads": {
      "post": {
        "tags": [
          "exercises"
        ],
        "summary": "Create Video Upload",
        "operationId": "create_video_upload_api_v1_exercises_video_uploads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseVideoUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/exercises/video-uploads/{media_id}/confirm": {
      "post": {
        "tags": [
          "exercises"
        ],
        "summary": "Confirm Video Upload",
        "operationId": "confirm_video_upload_api_v1_exercises_video_uploads__media_id__confirm_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseVideoConfirmResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exercises/video-uploads/{media_id}/retry-stream": {
      "post": {
        "tags": [
          "exercises"
        ],
        "summary": "Retry Video Stream",
        "operationId": "retry_video_stream_api_v1_exercises_video_uploads__media_id__retry_stream_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseVideoConfirmResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exercises/thumbnail-uploads": {
      "post": {
        "tags": [
          "exercises"
        ],
        "summary": "Create Thumbnail Upload",
        "operationId": "create_thumbnail_upload_api_v1_exercises_thumbnail_uploads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseThumbnailUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/exercises/thumbnail-uploads/{media_id}/confirm": {
      "post": {
        "tags": [
          "exercises"
        ],
        "summary": "Confirm Thumbnail Upload",
        "operationId": "confirm_thumbnail_upload_api_v1_exercises_thumbnail_uploads__media_id__confirm_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Media Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exercises/{exercise_id}": {
      "patch": {
        "tags": [
          "exercises"
        ],
        "summary": "Update Exercise",
        "operationId": "update_exercise_api_v1_exercises__exercise_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exercise Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExercisePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exercises/{exercise_id}/video/read-url": {
      "post": {
        "tags": [
          "exercises"
        ],
        "summary": "Create Exercise Video Read Url",
        "operationId": "create_exercise_video_read_url_api_v1_exercises__exercise_id__video_read_url_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exercise Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaReadUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exercises/{exercise_id}/video/stream.m3u8": {
      "get": {
        "tags": [
          "exercises"
        ],
        "summary": "Get Exercise Video Stream Manifest",
        "operationId": "get_exercise_video_stream_manifest_api_v1_exercises__exercise_id__video_stream_m3u8_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exercise Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exercises/{exercise_id}/thumbnail/read-url": {
      "post": {
        "tags": [
          "exercises"
        ],
        "summary": "Create Exercise Thumbnail Read Url",
        "operationId": "create_exercise_thumbnail_read_url_api_v1_exercises__exercise_id__thumbnail_read_url_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exercise Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaReadUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmUploadResponse": {
        "properties": {
          "media_id": {
            "type": "string",
            "title": "Media Id"
          },
          "status": {
            "type": "string",
            "const": "READY",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "media_id",
          "status"
        ],
        "title": "ConfirmUploadResponse"
      },
...
      "CreateUploadResponse": {
        "properties": {
          "media_id": {
            "type": "string",
            "title": "Media Id"
          },
          "upload_url": {
            "type": "string",
            "title": "Upload Url"
          },
          "upload_headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Upload Headers"
          },
          "expires_in_seconds": {
            "type": "integer",
            "title": "Expires In Seconds"
          }
        },
        "type": "object",
        "required": [
          "media_id",
          "upload_url",
          "upload_headers",
          "expires_in_seconds"
        ],
        "title": "CreateUploadResponse"
      },
...
      "ExerciseCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "direction": {
            "type": "string",
            "enum": [
              "speed",
              "strength",
              "agility",
              "cardio"
            ],
            "title": "Direction"
          },
          "muscle_group": {
            "type": "string",
            "enum": [
              "Ноги",
              "Грудь",
              "Спина",
              "Плечи",
              "Руки",
              "Кор",
              "Всё тело"
            ],
            "title": "Muscle Group"
          },
          "instruction": {
            "type": "string",
            "maxLength": 4000,
            "title": "Instruction",
            "default": ""
          },
          "reference_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Url"
          },
          "video_platform": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "rutube",
                  "youtube",
                  "vk"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Platform"
          },
          "video_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Url"
          },
          "video_file_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Video File Url"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          }
        },
        "type": "object",
        "required": [
          "title",
          "direction",
          "muscle_group"
        ],
        "title": "ExerciseCreate"
      },
      "ExercisePatch": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "instruction": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Instruction"
          },
          "muscle_groups": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "Ноги",
                    "Грудь",
                    "Спина",
                    "Плечи",
                    "Руки",
                    "Кор",
                    "Всё тело"
                  ]
                },
                "type": "array",
                "maxItems": 7,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Muscle Groups"
          },
          "video_media_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 36,
                "minLength": 36
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Media Id"
          },
          "thumbnail_media_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 36,
                "minLength": 36
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Media Id"
          }
        },
        "type": "object",
        "title": "ExercisePatch"
      },
      "ExerciseResponse": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "direction": {
            "type": "string",
            "enum": [
              "speed",
              "strength",
              "agility",
              "cardio"
            ],
            "title": "Direction"
          },
          "muscle_group": {
            "type": "string",
            "enum": [
              "Ноги",
              "Грудь",
              "Спина",
              "Плечи",
              "Руки",
              "Кор",
              "Всё тело"
            ],
            "title": "Muscle Group"
          },
          "instruction": {
            "type": "string",
            "maxLength": 4000,
            "title": "Instruction",
            "default": ""
          },
          "reference_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Url"
          },
          "video_platform": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "rutube",
                  "youtube",
                  "vk"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Platform"
          },
          "video_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Url"
          },
          "video_file_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Video File Url"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "trainer_id": {
            "type": "string",
            "title": "Trainer Id"
          },
          "muscle_groups": {
            "items": {
              "type": "string",
              "enum": [
                "Ноги",
                "Грудь",
                "Спина",
                "Плечи",
                "Руки",
                "Кор",
                "Всё тело"
              ]
            },
            "type": "array",
            "title": "Muscle Groups"
          },
          "video_media_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Media Id"
          },
          "thumbnail_media_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Media Id"
          },
          "video_stream_status": {
            "type": "string",
            "enum": [
              "NONE",
              "PROCESSING",
              "READY",
              "FAILED"
            ],
            "title": "Video Stream Status",
            "default": "NONE"
          }
        },
        "type": "object",
        "required": [
          "title",
          "direction",
          "muscle_group",
          "id",
          "trainer_id",
          "muscle_groups"
        ],
        "title": "ExerciseResponse"
      },
      "ExerciseThumbnailUploadRequest": {
        "properties": {
          "content_type": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ],
            "title": "Content Type"
          },
          "content_length": {
            "type": "integer",
            "maximum": 5242880.0,
            "exclusiveMinimum": 0.0,
            "title": "Content Length"
          }
        },
        "type": "object",
        "required": [
          "content_type",
          "content_length"
        ],
        "title": "ExerciseThumbnailUploadRequest"
      },
      "ExerciseVideoConfirmResponse": {
        "properties": {
          "media_id": {
            "type": "string",
            "title": "Media Id"
          },
          "status": {
            "type": "string",
            "const": "READY",
            "title": "Status"
          },
          "stream_status": {
            "type": "string",
            "enum": [
              "NONE",
              "PROCESSING",
              "READY",
              "FAILED"
            ],
            "title": "Stream Status"
          }
        },
        "type": "object",
        "required": [
          "media_id",
          "status",
          "stream_status"
        ],
        "title": "ExerciseVideoConfirmResponse"
      },
      "ExerciseVideoUploadRequest": {
        "properties": {
          "content_type": {
            "type": "string",
            "enum": [
              "video/mp4",
              "video/webm",
              "video/quicktime"
            ],
            "title": "Content Type"
          },
          "content_length": {
            "type": "integer",
            "maximum": 209715200.0,
            "exclusiveMinimum": 0.0,
            "title": "Content Length"
          }
        },
        "type": "object",
        "required": [
          "content_type",
          "content_length"
        ],
        "title": "ExerciseVideoUploadRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MediaReadUrlResponse": {
        "properties": {
          "media_id": {
            "type": "string",
            "title": "Media Id"
          },
          "read_url": {
            "type": "string",
            "title": "Read Url"
          },
          "expires_in_seconds": {
            "type": "integer",
            "title": "Expires In Seconds"
          }
        },
        "type": "object",
        "required": [
          "media_id",
          "read_url",
          "expires_in_seconds"
        ],
        "title": "MediaReadUrlResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}